requirements.txt
albumentations==1.3.1
huggingface-hub==0.16.4
jsonargparse==4.27.1
lightning==2.1.0
lightning-bolts==0.3.4
lightning-utilities==0.9.0
lit==16.0.6
omegaconf==2.3.0
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
pandas==1.3.3
Pillow==9.4.0
prettytable==3.9.0
pyparsing==3.0.9
pytorch-lightning==2.1.3
scikit-image==0.19.3
scikit-learn==1.3.0
scipy==1.13.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
timm==0.9.12
tokenizers==0.14.1
torch==2.0.1
torch-tb-profiler==0.4.3
torchdata==0.6.1
torchinfo==1.8.0
torchmetrics==1.2.1
torchsummary==1.5.1
torchtext==0.15.2
torchvision==0.15.2
tqdm==4.66.1

